move function to ./functions directory
